requirements.txt
Django==3.2
mecab-python3==1.0.4
ipadic==1.0.0
